feat: add on_async_action_completion column to webhook table (#4062)
We want to send webhook when async actions complete. Adding a column in
the db to store the setting value
<!-- Summary by @propel-code-bot -->

---

This PR adds a new boolean column `on_async_action_completion` to the
`_nango_external_webhooks` table with a default value of true. This
change enables the system to track whether webhooks should be sent when
asynchronous actions are completed.

*This summary was automatically generated by @propel-code-bot*