fix(webapp): delete button not working (#4060)
<!-- Describe the problem and your solution --> 
### Description
The delete button was stopped working at some point... This PR fixes
that and another bug, where you could trigger the modal even when the
button was blocked, by clicking on the tooltip.
You can see the bugs happening in staging or prod:
1. Delete button not working for any environment.
2. For the prod environment, we disable the button, but hovering over it
and clicking on the tooltip will trigger the modal (ironically).

![image](https://github.com/user-attachments/assets/8f8bf8ab-9911-49be-865f-434421f29583)
> Note: this doesn't make deleting the prod environment possible as it's
protected by the backend.

This PR fixes both problems.

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

<!-- Summary by @propel-code-bot -->

---

Removes an incorrect `asChild` prop that blocked the dialog from opening
and disables pointer events on the tooltip wrapper so the modal can’t be
triggered when the button is disabled.

*This summary was automatically generated by @propel-code-bot*