fix(proxy): do not append querystring on top of url (#4084)
## Changes

- fix(proxy): do not append querystring on top of url
When I upgraded express req.params was now incorrect due to change of
implementation, so I used req.originalUrl without realising it would
also contain the query params resulting in being set twice in the URL.
The good thing is that it simplifies the code and remove those weird old
school imports
<!-- Summary by @propel-code-bot -->

---

This PR updates the proxy controller to avoid appending the query string
twice in proxied requests. It removes unused legacy imports and
simplifies the construction of the endpoint URL by relying solely on
req.originalUrl, which already includes the query string.

*This summary was automatically generated by @propel-code-bot*