chore(billing): add dimensions to billing events
Adding prooperties like environmentId, connectionId, etc... to billing
events.
Even though those properties are not being used for billable metrics
they might be used in the future and/or to display more granular data to
customers
Note: for billable_connections, since we are already sending the
aggregated value, we cannot easily split it to add dimensions. I
therefore haven't added any dimension to it