chore: add group max concurrency to tasks table
The groups table is going to be removed since its sole purpose was to be
able to lock at the group level in `dequeue`. Since it doesn't allow for
a correct handling of max concurrency we are now using a advisory lock
and the groups table is not required anymore. I will remove stop using
the groups table in next PR and will remove the table eventually