chore: use github app to push release commit
In order to satisfies audit we need to remove branch protection bypass
for repo admins.
With this change the git commit for the changelog release commit will be
done via a github app and not via the admin token. We will then be able
to remove the branch protection bypass for admins