fix: add missing index for sync endpoint table (#4135)
We are joining on sync_config_id here:

https://github.com/NangoHQ/nango/blob/a093baae68861bce97b7ede53074eb9ab710c9da/packages/shared/lib/services/sync/config/config.service.ts#L707
but there were no index

<!-- Describe the problem and your solution --> 

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

<!-- Summary by @propel-code-bot -->

---

This PR introduces a database migration to add a btree index on the
sync_config_id column of the nango._nango_sync_endpoints table. The
index is intended to optimize join queries involving sync_config_id, as
highlighted in the referenced code location.

*This summary was automatically generated by @propel-code-bot*