fix: failed starting supervisor error when remote runner (#4141)
Do not initialize `Fleet.supervisor` if no node provider (which is the
case on enterprise setup with remote runners). Making start/stop no op
operations
 

<!-- Summary by @propel-code-bot -->

---

This PR updates the Fleet class to only initialize the Supervisor
instance if a node provider is provided. This change addresses errors
encountered when using remote runners in enterprise setups, where a node
provider may not exist.

*This summary was automatically generated by @propel-code-bot*