fix(shared): remove explicit setting of AWS credentials (#4189)
<!-- Describe the problem and your solution --> 
Removed the explicit setting of the AWS credentials when creating the S3
client as the credentials use the default names and will automatically
be detected by the AWS credentials provider. More details
[here](https://docs.aws.amazon.com/sdkref/latest/guide/standardized-credentials.html)
and for JavaScipt specifically,
[here](https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/setting-credentials-node.html#credchain)

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

<!-- Summary by @propel-code-bot -->

---

This PR updates the S3 client initialization in remote.service.ts to
remove the explicit specification of AWS access key and secret key,
relying instead on the AWS SDK's default credential provider chain.
Region selection logic is preserved with an enterprise-specific
fallback, resulting in a cleaner and more standardized setup.

*This summary was automatically generated by @propel-code-bot*