fix(db): add fields in plans for self-serve billing (#4227)
## Changes

-  add fields in plans for self-serve billing
    - `stripe_customer_id` -> to link in orb
    - `stripe_payment_id` -> to link in orb
    - `orb_customer_id` -> to know which customer has been linked to orb
    - `orb_subscription_id` -> to know which subscription a customer has
- `orb_future_plan` -> future (when we downgrade the next plan is kickin
at the start of the next billing period)
- `orb_future_plan_at` -> handy to not have to compute or if an admin
choose a different start period

<!-- Summary by @propel-code-bot -->

---

This PR introduces new columns to the 'plans' table to support
self-serve billing. Specifically, it adds Stripe and Orb identifiers, as
well as fields to manage future plan changes and their effective dates.
The migration is implemented through a Knex migration script.

*This summary was automatically generated by @propel-code-bot*