fix(plans): add auto_idle flag (#4234)
## Changes

Fixes
https://linear.app/nango/issue/NAN-3456/make-auto-idling-a-feature-flag

- add auto_idle flag
Not used yet but we'll replace most of the trial flag

<!-- Summary by @propel-code-bot -->

---

This PR introduces an 'auto_idle' boolean flag to the 'plans' table,
updates the corresponding type definitions, and ensures the new field is
reflected in seeded test data. The field is added with a default value
of true and is not yet used in business logic, but will later replace
most uses of the trial flag.

*This summary was automatically generated by @propel-code-bot*