chore: add last_execution_at column to connection table
we want to add a new metric to automate the billing of some of the
legacy scale plan. The metric is "number of connections with at least a
script execution during the period". There is currently no source of
truth in the db regarding script executions (actions for example are not
tracked). In order to avoid sending an event for every single
execution, this commit is adding a last_execution_at column to the
connection table that will be updated every time a script execution starts